requirements.txt
numpy
pandas
scikit-learn>=1.3
scipy
joblib
pyyaml
tqdm
torch>=2.0
sentence-transformers>=2.6
tf-keras